WIP – incorporate logger amendmends from experiences with Kazana
https://github.com/eHealthAfrica/kazana-server/tree/master/lib/plugins
