+++ lib/plugins/index.js
@@ -13,7 +13,7 @@ function registerPlugins (server, config, callback) {
     require('lout')
   ]
   var localPlugins = [
-    require('./log'),
+    require('./logger'),
     require('./maybe-force-gzip'),
     require('./public')
   ].map(function (register) {
